attacks/cleverhans/requirements.txt
tensorflow==2.11.0
tensorflow-datasets==4.8.2
cleverhans==4.0.0
functions-framework==3.0.0
google-cloud-storage==2.7.0